Workstation makes better use of vertical screen real estate https://www.pivotaltracker.com/story/show/186045980
