+++ gui/src/main/java/io/xj/gui/controllers/MainPaneBottomController.java
@@ -9,14 +9,12 @@ import jakarta.annotation.Nullable;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.application.Platform;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.css.PseudoClass;
-import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.control.*;
+import javafx.scene.control.ListCell;
+import javafx.scene.control.ListView;
 import javafx.scene.layout.VBox;
 import javafx.util.Duration;
 import org.springframework.beans.factory.annotation.Value;
@@ -29,8 +27,6 @@ import java.util.concurrent.LinkedBlockingDeque;
 
 @Service
 public class MainPaneBottomController extends VBox implements ReadyAfterBootController {
-  final BooleanProperty logsTailing = new SimpleBooleanProperty(true);
-  final BooleanProperty logsVisible = new SimpleBooleanProperty(false);
   final Integer refreshRateSeconds;
   final LabService labService;
   final LogQueue logQueue;
@@ -41,22 +37,10 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
   final static PseudoClass warn = PseudoClass.getPseudoClass("warn");
   static final int LOG_LIST_VIEW_HEIGHT = 368;
   static final int MAX_ENTRIES = 10_000;
-  final ModalLabConnectionController modalLabConnectionController;
+  final MainMenuController mainMenuController;
 
   @Nullable
   Timeline refresh;
-
-  @FXML
-  public Button buttonLab;
-
-  @FXML
-  public Label labelLabStatus;
-
-  @FXML
-  public ToggleButton toggleShowLogs;
-
-  @FXML
-  public ToggleButton toggleTailLogs;
 
   @FXML
   protected ListView<MainPaneBottomController.LogRecord> logListView;
@@ -65,11 +49,11 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
   public MainPaneBottomController(
     @Value("${gui.logs.refresh.seconds}") Integer refreshRateSeconds,
     LabService labService,
-    ModalLabConnectionController modalLabConnectionController
+    MainMenuController mainMenuController
   ) {
     this.refreshRateSeconds = refreshRateSeconds;
     this.labService = labService;
-    this.modalLabConnectionController = modalLabConnectionController;
+    this.mainMenuController = mainMenuController;
     logQueue = new LogQueue();
 
     // bind to the log appender
@@ -78,8 +62,6 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
 
   @Override
   public void onStageReady() {
-    labelLabStatus.textProperty().bind(labService.statusProperty().map(Enum::toString));
-
     refresh = new Timeline(
       new KeyFrame(
         Duration.seconds(1),
@@ -90,7 +72,7 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
             logItems.remove(0, logItems.size() - MAX_ENTRIES);
           }
 
-          if (logsTailing.get()) {
+          if (mainMenuController.logsTailingProperty().get()) {
             logListView.scrollTo(logItems.size());
           }
         }
@@ -127,13 +109,10 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
     });
     logListView.setItems(logItems);
 
-    toggleTailLogs.selectedProperty().bindBidirectional(logsTailing);
-    toggleShowLogs.selectedProperty().bindBidirectional(logsVisible);
-    logListView.visibleProperty().bind(logsVisible);
-    toggleTailLogs.visibleProperty().bind(logsVisible);
-    logListView.minHeightProperty().bind(logsVisible.map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
-    logListView.prefHeightProperty().bind(logsVisible.map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
-    logListView.maxHeightProperty().bind(logsVisible.map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
+    logListView.visibleProperty().bind(mainMenuController.logsVisibleProperty());
+    logListView.minHeightProperty().bind(mainMenuController.logsVisibleProperty().map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
+    logListView.prefHeightProperty().bind(mainMenuController.logsVisibleProperty().map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
+    logListView.maxHeightProperty().bind(mainMenuController.logsVisibleProperty().map((v) -> v ? LOG_LIST_VIEW_HEIGHT : 0));
   }
 
   @Override
@@ -150,11 +129,6 @@ public class MainPaneBottomController extends VBox implements ReadyAfterBootCont
       } catch (Exception e) {
         // no op
       }
-  }
-
-  @FXML
-  public void handleButtonLabPressed(ActionEvent ignored) {
-    modalLabConnectionController.launchModal();
   }
 
   static class LogQueue {
