+++ gui/src/main/resources/styles/dark-theme.css
@@ -83,7 +83,7 @@
   -fx-table-cell-border-color: transparent;
 }
 
-.bottom-pane-status-container, .top-pane-status-container {
+.top-pane-status-container {
   -fx-background-color: #494949;
 }
