+++ gui/src/main/resources/styles/default-theme.css
@@ -36,10 +36,6 @@
   -fx-text-fill: red;
 }
 
-.bottom-pane-status-container {
-  -fx-background-color: #c0c0c0;
-}
-
 .top-pane-status-container {
   -fx-background-color: #c9c9c9;
 }
